+++ CHANGES.rst
@@ -1,6 +1,24 @@
 Changelog
 =========
 
+3.3.3 (unreleased)
+------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- Move ``get_top_site_from_url`` out from here into ``Products.CMFPlone.utils``.
+  Deprecate old import.
+  [thet]
+
+Bug fixes:
+
+- *add item here*
+
+
 3.3.2 (unreleased)
 ------------------